feat: categorizedUsers
